account for connection issues

+++ src/pages/Play.vue
@@ -2,6 +2,8 @@
 .play
   Cover(v-if="!audioContext")
     h1 plz turn OFFFFF your content blockers and let me contRoL ur aud.io!
+  Cover(v-if="!connected")
+    h1 overcoming the teCHnical boundarIEs and CONNECTING...
 </template>
 
 <script>
@@ -16,13 +18,27 @@ export default {
     }
     this.sineInstrument = new SineInstrument(this.audioContext);
   },
+  data() {
+    return {
+      connected: false
+    }
+  },
   sockets: {
+    connect() {
+      this.connected = true;
+    },
+    disconnect() {
+      this.connected = false;
+      this.kill();
+    },
     play(options) {
       this.setPlayingInstrument(this.sineInstrument);
       this.playingInstrument.play(options);
     },
     kill(options) {
-      this.playingInstrument.kill(options);
+      if (this.playingInstrument) {
+        this.playingInstrument.kill(options);
+      }
     }
   },
   computed: {
